factor out contract and has-tx-object
