+++ src/index.js
@@ -2,135 +2,48 @@ const abi = require('ethjs-abi'); // eslint-disable-line
 const keccak256 = require('js-sha3').keccak_256; // eslint-disable-line
 const EthFilter = require('ethjs-filter'); // eslint-disable-line
 const getKeys = require('ethjs-util').getKeys; // eslint-disable-line
-const arrayContainsArray = require('ethjs-util').arrayContainsArray;
+const Contract = require('./contract');
+const hasTransactionObject = require('./has-tx-object');
 
-function hasTransactionObject(args) {
-  const txObjectProperties = ['from', 'to', 'data', 'value', 'gasPrice', 'gas'];
-  if (typeof args === 'object' && Array.isArray(args) === true && args.length > 0) {
-    if (typeof args[args.length - 1] === 'object'
-      && (Object.keys(args[args.length - 1]).length === 0
-      || arrayContainsArray(Object.keys(args[args.length - 1]), txObjectProperties, true))) {
-      return true;
-    }
-  }
+module.exports = EthContract;
 
-  return false;
-}
-
-function getConstructorFromABI(contractABI) {
-  return contractABI.filter((json) => (json.type === 'constructor'))[0];
-}
-
-function getCallableMethodsFromABI(contractABI) {
-  return contractABI.filter((json) => ((json.type === 'function' || json.type === 'event') && json.name.length > 0));
-}
-
-function contractFactory(query) {
-  return function ContractFactory(contractABI, contractBytecode, contractDefaultTxObject) {
+function EthContract(query) {
+  return function contractFactory(contractABI, contractBytecode, contractDefaultTxObject) {
+    // validate params
     if (!Array.isArray(contractABI)) { throw new Error(`[ethjs-contract] Contract ABI must be type Array, got type ${typeof contractABI}`); }
     if (typeof contractBytecode !== 'undefined' && typeof contractBytecode !== 'string') { throw new Error(`[ethjs-contract] Contract bytecode must be type String, got type ${typeof contractBytecode}`); }
     if (typeof contractDefaultTxObject !== 'undefined' && typeof contractDefaultTxObject !== 'object') { throw new Error(`[ethjs-contract] Contract default tx object must be type Object, got type ${typeof contractABI}`); }
 
+    // build contract object
     const output = {};
-    output.at = function atContract(address) {
-      function Contract() {
-        const self = this;
-        self.abi = contractABI || [];
-        self.query = query;
-        self.address = address || '0x';
-        self.bytecode = contractBytecode || '0x';
-        self.defaultTxObject = contractDefaultTxObject || {};
-        self.filters = new EthFilter(query);
-
-        getCallableMethodsFromABI(contractABI).forEach((methodObject) => {
-          self[methodObject.name] = function contractMethod() { // eslint-disable-line
-            var queryMethod = 'call'; // eslint-disable-line
-            var providedTxObject = {}; // eslint-disable-line
-            var methodCallback; // eslint-disable-line
-            const methodArgs = [].slice.call(arguments); // eslint-disable-line
-            if (typeof methodArgs[methodArgs.length - 1] === 'function') {
-              methodCallback = methodArgs.pop();
-            }
-
-            if (methodObject.type === 'function') {
-              const prom = new Promise((resolve, reject) => {
-                function newMethodCallback(callbackError, callbackResult) {
-                  if (queryMethod === 'call' && !callbackError) {
-                    try {
-                      const decodedMethodResult = abi.decodeMethod(methodObject, callbackResult);
-
-                      resolve(decodedMethodResult);
-                    } catch (decodeFormattingError) {
-                      const decodingError = new Error(`[ethjs-contract] while formatting incoming raw call data ${JSON.stringify(callbackResult)} ${decodeFormattingError}`);
-
-                      reject(decodingError);
-                    }
-                  } else if (queryMethod === 'sendTransaction' && !callbackError) {
-                    resolve(callbackResult);
-                  } else {
-                    reject(callbackError);
-                  }
-                }
-
-                if (hasTransactionObject(methodArgs)) providedTxObject = methodArgs.pop();
-                const methodTxObject = Object.assign({},
-                  self.defaultTxObject,
-                  providedTxObject, {
-                    to: self.address,
-                  });
-                methodTxObject.data = abi.encodeMethod(methodObject, methodArgs);
-
-                if (methodObject.constant === false) {
-                  queryMethod = 'sendTransaction';
-                }
-
-                query[queryMethod](methodTxObject, newMethodCallback);
-              });
-
-              if (methodCallback) {
-                prom.then(res => methodCallback(null, res)).catch(err => methodCallback(err, null));
-              }
-
-              return methodCallback ? null : prom;
-            } else if (methodObject.type === 'event') {
-              const filterInputTypes = getKeys(methodObject.inputs, 'type', false);
-              const filterTopic = `0x${keccak256(`${methodObject.name}(${filterInputTypes.join(',')})`)}`;
-              const filterTopcis = [filterTopic];
-              const argsObject = Object.assign({}, methodArgs[0]) || {};
-
-              return new self.filters.Filter(Object.assign({}, argsObject, {
-                decoder: (logData) => abi.decodeEvent(methodObject, logData, filterTopcis),
-                defaultFilterObject: Object.assign({}, (methodArgs[0] || {}), {
-                  to: self.address,
-                  topics: filterTopcis,
-                }),
-              }));
-            }
-          };
-        });
-      }
-
-      return new Contract();
+    output.at = function contractAtAddress(address) {
+      return new Contract({
+        address,
+        query,
+        contractBytecode,
+        contractDefaultTxObject,
+        contractABI,
+      });
     };
 
     output.new = function newContract() {
-      var providedTxObject = {}; // eslint-disable-line
-      var newMethodCallback = null; // eslint-disable-line
+      let providedTxObject = {}; // eslint-disable-line
+      let newMethodCallback = null; // eslint-disable-line
       const newMethodArgs = [].slice.call(arguments); // eslint-disable-line
       if (typeof newMethodArgs[newMethodArgs.length - 1] === 'function') newMethodCallback = newMethodArgs.pop();
       if (hasTransactionObject(newMethodArgs)) providedTxObject = newMethodArgs.pop();
-      const constructMethod = getConstructorFromABI(contractABI);
+      const constructorMethod = getConstructorFromABI(contractABI);
       const assembleTxObject = Object.assign({}, contractDefaultTxObject, providedTxObject);
 
-      // if contract bytecode was predefined
+      // set contract deploy bytecode
       if (contractBytecode) {
         assembleTxObject.data = contractBytecode;
       }
 
-      // if constructor bytecode
-      if (constructMethod) {
-        const constructBytecode = abi.encodeParams(getKeys(constructMethod.inputs, 'type'), newMethodArgs).substring(2); // eslint-disable-line
-        assembleTxObject.data = `${assembleTxObject.data}${constructBytecode}`;
+      // append encoded constructor arguments
+      if (constructorMethod) {
+        const constructorBytecode = abi.encodeParams(getKeys(constructorMethod.inputs, 'type'), newMethodArgs).substring(2); // eslint-disable-line
+        assembleTxObject.data = `${assembleTxObject.data}${constructorBytecode}`;
       }
 
       return newMethodCallback ? query.sendTransaction(assembleTxObject, newMethodCallback) : query.sendTransaction(assembleTxObject);
@@ -140,8 +53,6 @@ function contractFactory(query) {
   };
 }
 
-function EthContract(query) {
-  return contractFactory(query);
+function getConstructorFromABI(contractABI) {
+  return contractABI.filter((json) => (json.type === 'constructor'))[0];
 }
-
-module.exports = EthContract;
